datamart/requirements.txt
Flask
Flask-Cors
Flask-JWT-Extended
Flask-RESTful
psycopg2-binary
uWSGI
openpyxl
pandas
